src/backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile src/backend/requirements.in
#
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
beautifulsoup4==4.12.2
    # via bs4
bs4==0.0.1
    # via -r src/backend/requirements.in
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
django==4.2.7
    # via
    #   -r src/backend/requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
django-cors-headers==4.3.1
    # via -r src/backend/requirements.in
django-filter==23.4
    # via -r src/backend/requirements.in
djangorestframework==3.14.0
    # via -r src/backend/requirements.in
idna==3.5
    # via requests
markdown==3.5.1
    # via -r src/backend/requirements.in
packaging==23.2
    # via build
pip-tools==7.3.0
    # via -r src/backend/requirements.in
psycopg2-binary==2.9.9
    # via -r src/backend/requirements.in
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via -r src/backend/requirements.in
pytz==2023.3.post1
    # via djangorestframework
requests==2.31.0
    # via -r src/backend/requirements.in
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.8.0
    # via asgiref
urllib3==2.1.0
    # via requests
wheel==0.41.3
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
